refactor(hwp5): split decoder/section.rs into submodules (E7 #4)
section.rs(4,602줄)를 section/{model,gso}.rs + mod.rs 로 분할.
- model.rs: pub(crate) 데이터 모델 ~25 타입(Hwp5Paragraph, Hwp5Control
  + 모든 Hwp5*Control, Hwp5Table 계열, SectionResult, Hwp5PageBorderFill 등)
- gso.rs: GSO 빌더(InlineGsoContext, GsoChildBuilder/GsoGroupBuilder 등)
- mod.rs: 파서 코어(BodyTextParserState) + table context + 41 테스트 유지
mod.rs 가 pub use 로 재노출 → crate::decoder::section::Type 경로 불변.
table 파서 로직은 BodyTextParserState 와 얽혀 있어 보존(추출 안 함).

순수 기계적 분할 — 동작·공개 API byte-동일. make ci 2536 통과.
